requirements.txt
flask
pillow
redis
sympy